Document the extensibility surface and split out ProviderToolAnnotations
Follow-up to 948c5c7, which added run-time tool contribution but left the
public surface undocumented.

- Full XML docs on every public member of McpExtensionRegistry, matching the
  style ExtensionProtocol already set. They record the reasoning a reader
  would otherwise have to reconstruct: Register returns a string rather than
  throwing because a clash between third-party plug-ins must never take the
  server down; same-owner re-registration is an update, not a clash, so a
  plug-in that reloads need not unregister first; UnregisterByOwner is
  deliberately not atomic, because locking the registry across a shutdown
  would let a misbehaving plug-in stall every request thread; Transformers()
  breaks ties by owner so chain order does not vary with Rhino's plug-in load
  order; and Tools is a live view rather than a snapshot.
- XML docs on McpEndpoint.TryResolveTool, in particular why compiled tools are
  consulted first: a contributed tool can then never shadow a built-in one,
  which together with the reserved ext_ prefix closes both halves of the
  impersonation gap. Also why the registry is read live rather than cached —
  it is what makes a tool registered after start-up callable without a
  restart.
- ProviderToolAnnotations moved out of ProviderToolDescriptor.cs into its own
  file.
- Every <summary> tag now sits on its own line across Server/Extensibility/.

Also commits ProviderToolDescriptor.cs and ResultTransformer.cs, which were
untracked.

Build: 0 errors, 9 pre-existing upstream warnings.
